\docType{package}
\name{IgGeneUsage-package}
\alias{IgGeneUsage-package}
\alias{IgGeneUsage}

\title{
The R package IgGeneUsage
}

\description{
IgGeneUsage detects aberrant immunoglobulin (Ig) gene usage between adaptive 
immune repertoires that belong to different biological conditions using a 
probabilistic model which is analyzed computationally by Bayes inference.
}

\details{
This package contains functions for:
\enumerate{
\item differential Ig gene usage analysis (function DGU)
\item posterior predictive checks (part of results generated by function DGU)
\item leave-one-out cross validation (function LOO)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/snaketron/IgGeneUsage}
  \item Report bugs at \url{https://github.com/snaketron/IgGeneUsage/issues}
}
}

\author{

Authors and maintainers:
\itemize{
  \item Simo Kitanovski \email{simokitanovski@uni-due.de}
  (\href{https://orcid.org/0000-0003-2909-5376}{ORCID})
}
}
