% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgGeneUsage-package.R
\docType{package}
\name{IgGeneUsage-package}
\alias{IgGeneUsage-package}
\alias{IgGeneUsage}
\title{The 'IgGeneUsage' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
